requirements.txt
feedparser>=6.0.0
paho-mqtt>=2.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
